Task_1_3_1

+++ Task_1_3_1/src/test/java/ru/nsu/demidov/substring/SubStringFinderTest.java
@@ -36,7 +36,7 @@ public class SubStringFinderTest {
         long OCCURENCES = 1000000000000000000L;
         try (BufferedWriter writer = new BufferedWriter(new FileWriter(tempFile))) {
             StringBuilder sb = new StringBuilder();
-            for (int i = 0; i < OCCURENCES; i++) {
+            for (long i = 0; i < OCCURENCES; i++) {
                 sb.append("boobsieboobiesboobs");
             }
             writer.write(sb.toString());
